+++ src/app.rs
@@ -43,6 +43,18 @@ pub struct AppConfigFields {
     pub disable_click: bool,
 }
 
+/// For filtering out information
+pub struct DataFilters {
+    pub disk_filter: Option<Filter>,
+    pub temp_filter: Option<Filter>,
+}
+
+#[derive(Debug)]
+pub struct Filter {
+    pub is_list_ignored: bool,
+    pub list: Vec<regex::Regex>,
+}
+
 #[derive(TypedBuilder)]
 pub struct App {
     #[builder(default = false, setter(skip))]
@@ -99,6 +111,7 @@ pub struct App {
     pub widget_map: HashMap<u64, BottomWidget>,
     pub current_widget: BottomWidget,
     pub used_widgets: UsedWidgets,
+    pub filters: DataFilters,
 }
 
 impl App {
@@ -312,10 +325,9 @@ impl App {
 
     pub fn toggle_sort(&mut self) {
         match &self.current_widget.widget_type {
-            // FIXME: [REFACTOR] Remove these @'s if unneeded, they were an idea but they're ultimately useless for me here...?
-            widget_type @ BottomWidgetType::Proc | widget_type @ BottomWidgetType::ProcSort => {
+            BottomWidgetType::Proc | BottomWidgetType::ProcSort => {
                 let widget_id = self.current_widget.widget_id
-                    - match &widget_type {
+                    - match &self.current_widget.widget_type {
                         BottomWidgetType::Proc => 0,
                         BottomWidgetType::ProcSort => 2,
                         _ => 0,
@@ -348,9 +360,9 @@ impl App {
 
     pub fn invert_sort(&mut self) {
         match &self.current_widget.widget_type {
-            widget_type @ BottomWidgetType::Proc | widget_type @ BottomWidgetType::ProcSort => {
+            BottomWidgetType::Proc | BottomWidgetType::ProcSort => {
                 let widget_id = self.current_widget.widget_id
-                    - match &widget_type {
+                    - match &self.current_widget.widget_type {
                         BottomWidgetType::Proc => 0,
                         BottomWidgetType::ProcSort => 2,
                         _ => 0,
@@ -1571,9 +1583,9 @@ impl App {
                     }
                 }
                 WidgetDirection::Down => match &self.current_widget.widget_type {
-                    proc_type @ BottomWidgetType::Proc | proc_type @ BottomWidgetType::ProcSort => {
+                    BottomWidgetType::Proc | BottomWidgetType::ProcSort => {
                         let widget_id = self.current_widget.widget_id
-                            - match proc_type {
+                            - match &self.current_widget.widget_type {
                                 BottomWidgetType::ProcSort => 2,
                                 _ => 0,
                             };